auth-service/requirements.txt
# requirements.txt - Add these
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
databases==0.8.0
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.1.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
httpx==0.25.0  # For GitHub API calls